+++ spec/features/staff/users_can_edit_an_activity_spec.rb
@@ -107,14 +107,14 @@ RSpec.feature "Users can edit an activity" do
       context "when the activity is complete" do
         it "editing and saving a step returns the user to the activity page details tab" do
           activity = create(:fund_activity, organisation: user.organisation)
-          identifier = "AB-CDE-1234"
+          description = "Some new text for the description field."
           visit organisation_activity_details_path(activity.organisation, activity)
 
-          within(".identifier") do
+          within(".description") do
             click_on(t("default.link.edit"))
           end
 
-          fill_in "activity[delivery_partner_identifier]", with: identifier
+          fill_in "activity[description]", with: description
           click_button t("form.button.activity.submit")
 
           expect(page).to have_content t("action.fund.update.success")
@@ -149,20 +149,20 @@ RSpec.feature "Users can edit an activity" do
 
         it "tracks activity updates with public_activity" do
           activity = create(:fund_activity, organisation: user.organisation)
-          identifier = "AB-CDE-1234"
+          description = "Some new text for the description field."
           PublicActivity.with_tracking do
             visit organisation_activity_details_path(activity.organisation, activity)
 
-            within(".identifier") do
+            within(".description") do
               click_on(t("default.link.edit"))
             end
 
-            fill_in "activity[delivery_partner_identifier]", with: identifier
+            fill_in "activity[description]", with: description
             click_button t("form.button.activity.submit")
 
             # grab the most recently created auditable_event
             auditable_events = PublicActivity::Activity.where(trackable_id: activity.id).order("created_at ASC")
-            expect(auditable_events.first.key).to eq "activity.update.identifier"
+            expect(auditable_events.first.key).to eq "activity.update.purpose"
             expect(auditable_events.first.owner_id).to eq user.id
             expect(auditable_events.first.trackable_id).to eq activity.id
           end
@@ -189,14 +189,13 @@ RSpec.feature "Users can edit an activity" do
         end
 
         context "when the activity only has an identifier" do
-          it "shows edit link on the identifier, and add link on only the next step" do
+          it "only shows the add link on the next step" do
             activity = create(:fund_activity, :at_purpose_step, organisation: user.organisation)
 
             visit organisation_activity_details_path(activity.organisation, activity)
 
-            # Click the first edit link that opens the form on step 1
             within(".identifier") do
-              expect(page).to have_content(t("default.link.edit"))
+              expect(page).not_to have_content(t("default.link.edit"))
             end
 
             within(".title") do
@@ -204,7 +203,7 @@ RSpec.feature "Users can edit an activity" do
             end
 
             within(".sector") do
-              expect(page).to_not have_content(t("default.link.add"))
+              expect(page).not_to have_content(t("default.link.add"))
             end
           end
         end
@@ -298,6 +297,34 @@ RSpec.feature "Users can edit an activity" do
         visit organisation_activity_details_path(activity.organisation, activity)
 
         expect(page).to_not have_content(t("summary.label.activity.publish_to_iati.label"))
+      end
+
+      scenario "the delivery partner identifier cannot be changed" do
+        activity = create(:project_activity, organisation: user.organisation)
+        _report = create(:report, state: :active, organisation: user.organisation, fund: activity.associated_fund)
+
+        visit organisation_activity_details_path(activity.organisation, activity)
+
+        within(".identifier") do
+          expect(page).not_to have_content(t("default.link.edit"))
+        end
+
+        # not even by visiting the URL directly
+        visit activity_step_path(activity, :identifier)
+        expect(page).not_to have_content("Enter your unique identifier")
+      end
+
+      context "when the project does not have a delivery partner identifier" do
+        scenario "the delivery partner identifier can be added" do
+          activity = create(:project_activity, :at_identifier_step, organisation: user.organisation)
+          _report = create(:report, state: :active, organisation: user.organisation, fund: activity.associated_fund)
+
+          visit organisation_activity_details_path(activity.organisation, activity)
+
+          within(".identifier") do
+            expect(page).to have_content(t("default.link.add"))
+          end
+        end
       end
 
       context "when the project does not have a RODA identifier" do
@@ -419,14 +446,20 @@ RSpec.feature "Users can edit an activity" do
 end
 
 def assert_all_edit_links_go_to_the_correct_form_step(activity:)
-  within(".identifier") do
-    click_on t("default.link.edit")
-    expect(page).to have_current_path(
-      activity_step_path(activity, :identifier)
-    )
+  if activity.delivery_partner_identifier.blank?
+    within(".identifier") do
+      click_on t("default.link.edit")
+      expect(page).to have_current_path(
+        activity_step_path(activity, :identifier)
+      )
+    end
+    click_on t("default.link.back")
+    click_on t("tabs.activity.details")
+  else
+    within(".identifier") do
+      expect(page).to_not have_link(t("default.link.edit"))
+    end
   end
-  click_on t("default.link.back")
-  click_on t("tabs.activity.details")
 
   within(".sector") do
     click_on(t("default.link.edit"))